Ufo y aliens

+++ src/edu/austral/prog2_2018c2/Level.java
@@ -3,16 +3,16 @@ package edu.austral.prog2_2018c2;
 public class Level {
 
     public int shields;
-    public double movementSpeed;
+    public int movementSpeed;
     public int num;
 
-    public Level(int number,int shields,double movementSpeed){
+    public Level(int number,int shields,int movementSpeed){
         this.shields = shields;
         this.movementSpeed = movementSpeed;
         num= number;
     }
 
-    public double getMovementSpeed() {
+    public int getMovementSpeed() {
         return movementSpeed;
     }
